shadow-contrast/requirements.txt
torch==1.8.1+cu111
torchvision==0.9.1+cu111
torchaudio==0.8.1
timm==0.6.7
einops==0.4.1
scikit-image==0.19.3
opencv-python==4.6.0.66
Pillow==9.2.0
rawpy==0.18.1
PyWavelets==1.3.0
numpy==1.21.6
scipy==1.7.3
pandas==1.3.5
python-dateutil==2.8.2
PyYAML==6.0
tqdm==4.64.1
tensorboardX==2.6
wandb==0.13.2
thop==0.1.1.post2209072238
lmdb==1.3.0
